+++ src/components/pages/UsersPage/View/index.js
@@ -1,62 +1,194 @@
 import React, { Fragment } from 'react'
-// import PropTypes from 'prop-types'
+import PropTypes from 'prop-types'
 import { withStyles } from 'material-ui/styles';
-import Typography from 'material-ui/Typography';
-import Grid from 'material-ui/Grid';
+// import Typography from 'material-ui/Typography';
+// import Grid from 'material-ui/Grid';
 
 import {
   styles,
   UsersPageView as PrismaCmsUsersPageView,
 } from '@prisma-cms/front/lib/modules/pages/UsersPage/View';
 
- 
-import {
-  Pagination,
-  UserLink,
-} from '@modxclub/ui';
+import Filters from "@prisma-cms/filters";
 
 import moment from 'moment';
+import { Typography } from 'material-ui';
 
- 
+
 export class UsersPageView extends PrismaCmsUsersPageView {
- 
 
-  static defaultProps = {
-    ...PrismaCmsUsersPageView.defaultProps,
-    columnData: [
+  static propTypes = {
+    ...PrismaCmsUsersPageView.propTypes,
+    filters: PropTypes.object,
+    setFilters: PropTypes.func,
+  };
+
+
+  // static defaultProps = {
+  //   ...PrismaCmsUsersPageView.defaultProps,
+  // };
+
+
+  renderFilters() {
+
+    const {
+      filters,
+      setFilters,
+    } = this.props;
+
+    return filters && setFilters ? <Filters
+      queryName="users"
+      filters={filters}
+      setFilters={setFilters}
+    /> : null;
+  }
+
+
+  getColumns() {
+
+    const {
+      UserLink,
+      ProjectLink,
+      Link,
+    } = this.context;
+
+    return [
       {
         id: 'username',
         numeric: false,
         disablePadding: false,
-        label: 'User',
+        label: 'Пользователь',
         renderer: (value, record) => {
- 
 
           return <UserLink
             user={record}
           />;
- 
-
         }
       },
       {
-        id: 'created_at',
+        id: 'createdAt',
         numeric: false,
         disablePadding: false,
-        label: 'Reg date',
+        label: 'Дата регистрации',
         renderer: (value, record) => {
- 
 
           return value ? moment(value).format("LL") : "";
-
         }
       },
-    ],
-  };
+      {
+        id: 'ProjectsCreated',
+        numeric: false,
+        disablePadding: false,
+        label: 'Создал проекты',
+        renderer: (value, record) => {
 
- 
+          const {
+            id: userId,
+            username,
+          } = record;
+
+          let output;
+
+          let max = 3;
+
+          if (value && value.length) {
+
+            const projects = value.map(n => {
+
+              const {
+                id,
+              } = n;
+
+              return <ProjectLink
+                key={id}
+                object={n}
+              />
+
+            });
+
+            const listProjects = projects.splice(0, max);
+
+            const moreProjects = projects.length;
+
+            output = <Fragment>
+
+              {listProjects.reduce((a, b) => [a, ", ", b])} {moreProjects ? <Link
+                to={`/projects?filters=%7B"CreatedBy"%3A%7B"username"%3A"${username}"%7D%7D`}
+              >
+                <Typography
+                  variant="caption"
+                >
+                  и еще {moreProjects}
+                </Typography>
+              </Link> : null}
+
+            </Fragment>
+
+          };
+
+          return output;
+        }
+      },
+      {
+        id: 'Projects',
+        numeric: false,
+        disablePadding: false,
+        label: 'Участвует в проектах',
+        renderer: (value, record) => {
+
+          const {
+            id: userId,
+            username,
+          } = record;
+
+          let output;
+
+          let max = 3;
+
+          if (value && value.length) {
+
+            const projects = value.map(n => {
+
+              const {
+                id,
+                Project,
+              } = n;
+
+              return <ProjectLink
+                key={id}
+                object={Project}
+              />
+
+            });
+
+            const listProjects = projects.splice(0, max);
+
+            const moreProjects = projects.length;
+
+            output = <Fragment>
+
+              {listProjects.reduce((a, b) => [a, ", ", b])} {moreProjects ? <Link
+                to={`/projects?filters=%7B"Members_some"%3A%7B"User"%3A%7B"username"%3A"${username}"%7D%7D%7D`}
+              >
+                <Typography
+                  variant="caption"
+                >
+                  и еще {moreProjects}
+                </Typography>
+              </Link> : null}
+
+            </Fragment>
+
+          };
+
+          return output;
+        }
+      },
+    ]
+  }
+
 
 }
 
 
-export default withStyles(styles)(props => <UsersPageView {...props}/>);
+export default withStyles(styles)(props => <UsersPageView {...props} />);